fix gramatical error, and refactor constants
