+++ .idea/modules/pluginClean.test.iml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id="pluginClean:test" external.linked.project.path="$MODULE_DIR$/../.." external.root.project.path="$MODULE_DIR$/../.." external.system.id="GRADLE" external.system.module.group="co.com.bancolombia.cleanArchitecture" external.system.module.type="sourceSet" external.system.module.version="1.5" type="JAVA_MODULE" version="4">
+<module external.linked.project.id="pluginClean:test" external.linked.project.path="$MODULE_DIR$/../.." external.root.project.path="$MODULE_DIR$/../.." external.system.id="GRADLE" external.system.module.group="co.com.bancolombia.cleanArchitecture" external.system.module.type="sourceSet" external.system.module.version="1.5.1" type="JAVA_MODULE" version="4">
   <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_11">
     <output-test url="file://$MODULE_DIR$/../../build/classes/java/test" />
     <exclude-output />
@@ -22,43 +22,7 @@
     <orderEntry type="module-library">
       <library>
         <CLASSES>
-          <root url="jar://$USER_HOME$/.gradle/wrapper/dists/gradle-5.6.3-bin/71q0cerxks9z090hthk8vq8r7/gradle-5.6.3/lib/kotlin-stdlib-1.3.41.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
           <root url="jar://$USER_HOME$/.gradle/wrapper/dists/gradle-5.6.3-bin/71q0cerxks9z090hthk8vq8r7/gradle-5.6.3/lib/kotlin-reflect-1.3.41.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$USER_HOME$/.gradle/wrapper/dists/gradle-5.6.3-bin/71q0cerxks9z090hthk8vq8r7/gradle-5.6.3/lib/groovy-all-1.3-2.5.4.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$USER_HOME$/.gradle/wrapper/dists/gradle-5.6.3-bin/71q0cerxks9z090hthk8vq8r7/gradle-5.6.3/lib/kotlin-stdlib-common-1.3.41.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$USER_HOME$/.gradle/wrapper/dists/gradle-5.6.3-bin/71q0cerxks9z090hthk8vq8r7/gradle-5.6.3/lib/kotlin-stdlib-jdk8-1.3.41.jar!/" />
         </CLASSES>
         <JAVADOC />
         <SOURCES />
@@ -77,7 +41,43 @@
     <orderEntry type="module-library">
       <library>
         <CLASSES>
+          <root url="jar://$USER_HOME$/.gradle/wrapper/dists/gradle-5.6.3-bin/71q0cerxks9z090hthk8vq8r7/gradle-5.6.3/lib/groovy-all-1.3-2.5.4.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
           <root url="jar://$USER_HOME$/.gradle/caches/5.6.3/generated-gradle-jars/gradle-api-5.6.3.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$USER_HOME$/.gradle/wrapper/dists/gradle-5.6.3-bin/71q0cerxks9z090hthk8vq8r7/gradle-5.6.3/lib/kotlin-stdlib-1.3.41.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$USER_HOME$/.gradle/wrapper/dists/gradle-5.6.3-bin/71q0cerxks9z090hthk8vq8r7/gradle-5.6.3/lib/kotlin-stdlib-common-1.3.41.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$USER_HOME$/.gradle/wrapper/dists/gradle-5.6.3-bin/71q0cerxks9z090hthk8vq8r7/gradle-5.6.3/lib/kotlin-stdlib-jdk8-1.3.41.jar!/" />
         </CLASSES>
         <JAVADOC />
         <SOURCES />
