+++ gradle.properties
@@ -1,2 +1,2 @@
 package=co.com.bancolombia
-systemProp.version=1.5
+systemProp.version=1.5.1
