+++ src/main/java/co/com/bancolombia/factory/EntryPointFactoryImpl.java
@@ -1,12 +1,12 @@
 package co.com.bancolombia.factory;
 
-
-import co.com.bancolombia.Constants;
+import co.com.bancolombia.templates.Constants;
 import co.com.bancolombia.Utils;
 import co.com.bancolombia.exceptions.CleanException;
 import co.com.bancolombia.models.Module;
 import co.com.bancolombia.models.entrypoints.ApiReactive;
 import co.com.bancolombia.models.entrypoints.ApiRest;
+import co.com.bancolombia.templates.EntryPointTemplate;
 
 import java.io.IOException;
 
@@ -15,7 +15,7 @@ public class EntryPointFactoryImpl implements ModuleFactory {
     @Override
     public Module makeDrivenAdapter(int codeEntryPoint) throws IOException, CleanException {
         Module entryPoint = null;
-        if (Constants.getNameEntryPoint(codeEntryPoint) == null) {
+        if (EntryPointTemplate.getNameEntryPoint(codeEntryPoint) == null) {
             throw new IllegalArgumentException("Entry Point not is available (1 -> API REST, 2 -> API REACTIVE)");
         }
