+++ src/main/java/co/com/bancolombia/models/drivenadapters/JPADrivenAdapter.java
@@ -1,47 +1,48 @@
 package co.com.bancolombia.models.drivenadapters;
 
-import co.com.bancolombia.Constants;
 import co.com.bancolombia.models.Module;
+import co.com.bancolombia.templates.DrivenAdapterTemplate;
+import co.com.bancolombia.templates.HelperTemplate;
 
 public class JPADrivenAdapter extends Module {
     @Override
     public String getClassNameModule() {
-        return Constants.JPA_REPOSITORY_CLASS;
+        return DrivenAdapterTemplate.JPA_REPOSITORY_CLASS;
     }
 
     @Override
     public String getModuleClassContent() {
-        return Constants.getJPARepositoryClassContent(super.getPackageName().concat(".").concat(super.getModulePackage()));
+        return DrivenAdapterTemplate.getJPARepositoryClassContent(super.getPackageName().concat(".").concat(super.getModulePackage()));
     }
 
     @Override
     public String getInterfaceNameModule() {
-        return Constants.JPA_REPOSITORY_INTERFACE;
+        return DrivenAdapterTemplate.JPA_REPOSITORY_INTERFACE;
     }
 
     @Override
     public String getModuleInterfaceContent() {
-        return Constants.getJPARepositoryInterfaceContent(super.getPackageName().concat(".").concat(super.getModulePackage()));
+        return DrivenAdapterTemplate.getJPARepositoryInterfaceContent(super.getPackageName().concat(".").concat(super.getModulePackage()));
     }
 
     @Override
     public String getBuildGradleModule() {
-        return  Constants.getBuildGradleHelperJPARepository();
+        return  HelperTemplate.getBuildGradleHelperJPARepository();
     }
 
     @Override
     public String getHelperModuleClassContent() {
-        return  Constants.getHelperJPARepositoryClassContent(super.getPackageName().concat(".").concat(super.getHelperPackage()));
+        return  HelperTemplate.getHelperJPARepositoryClassContent(super.getPackageName().concat(".").concat(super.getHelperPackage()));
     }
 
     @Override
     public String getBuildGradleContentModule() {
-        return Constants.getBuildGradleJPARepository();
+        return DrivenAdapterTemplate.getBuildGradleJPARepository();
     }
 
     @Override
     public String getSettingsGradleModule() {
-        return Constants.getSettingsJPARepositoryContent().concat(Constants.getSettingsHelperJPAContent());
+        return DrivenAdapterTemplate.getSettingsJPARepositoryContent().concat(HelperTemplate.getSettingsHelperJPAContent());
     }
 
     @Override
