+++ src/main/java/co/com/bancolombia/task/GenerateModelTask.java
@@ -1,7 +1,9 @@
 package co.com.bancolombia.task;
 
-import co.com.bancolombia.Constants;
+import co.com.bancolombia.templates.Constants;
 import co.com.bancolombia.Utils;
+import co.com.bancolombia.templates.ModelTemplate;
+import co.com.bancolombia.templates.PluginTemplate;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.tasks.TaskAction;
@@ -27,16 +29,16 @@ public class GenerateModelTask extends DefaultTask {
         logger.lifecycle("Project  Package: {}", packageName);
         packageName = packageName.replaceAll("\\.", "\\/");
         logger.lifecycle("Model Name: {}", modelName);
-        logger.lifecycle("Generating Childs Dirs");
+        logger.lifecycle(PluginTemplate.GENERATING_CHILDS_DIRS);
         getProject().mkdir(Constants.DOMAIN.concat("/").concat(Constants.MODEL).concat("/").concat(Constants.MAIN_JAVA).concat("/").concat(packageName).concat("/").concat(Constants.MODEL).concat("/").concat(Utils.decapitalize(modelName)).concat("/").concat(Constants.GATEWAYS));
         getProject().mkdir(Constants.DOMAIN.concat("/").concat(Constants.MODEL).concat("/").concat(Constants.TEST_JAVA).concat("/").concat(packageName).concat("/").concat(Constants.MODEL).concat("/").concat(Utils.decapitalize(modelName)));
 
-        logger.lifecycle("Generated Childs Dirs");
+        logger.lifecycle(PluginTemplate.GENERATED_CHILDS_DIRS);
 
-        logger.lifecycle("Writing in Files");
-        Utils.writeString(getProject(), Constants.DOMAIN.concat("/").concat(Constants.MODEL).concat("/").concat(Constants.MAIN_JAVA).concat("/").concat(packageName).concat("/").concat(Constants.MODEL).concat("/").concat(Utils.decapitalize(modelName)).concat("/").concat(Constants.GATEWAYS).concat("/").concat(Utils.capitalize(modelName) + Constants.REPOSITORY + Constants.JAVA_EXTENSION), Constants.getInterfaceModel(modelName, packageName));
-        Utils.writeString(getProject(), Constants.DOMAIN.concat("/").concat(Constants.MODEL).concat("/").concat(Constants.MAIN_JAVA).concat("/").concat(packageName).concat("/").concat(Constants.MODEL).concat("/").concat(Utils.decapitalize(modelName)).concat("/").concat(Utils.capitalize(modelName) + Constants.JAVA_EXTENSION), Constants.getModel(modelName, packageName));
-        logger.lifecycle("Writed in Files");
+        logger.lifecycle(PluginTemplate.GENERATING_FILES);
+        Utils.writeString(getProject(), Constants.DOMAIN.concat("/").concat(Constants.MODEL).concat("/").concat(Constants.MAIN_JAVA).concat("/").concat(packageName).concat("/").concat(Constants.MODEL).concat("/").concat(Utils.decapitalize(modelName)).concat("/").concat(Constants.GATEWAYS).concat("/").concat(Utils.capitalize(modelName) + Constants.REPOSITORY + Constants.JAVA_EXTENSION), ModelTemplate.getInterfaceModel(modelName, packageName));
+        Utils.writeString(getProject(), Constants.DOMAIN.concat("/").concat(Constants.MODEL).concat("/").concat(Constants.MAIN_JAVA).concat("/").concat(packageName).concat("/").concat(Constants.MODEL).concat("/").concat(Utils.decapitalize(modelName)).concat("/").concat(Utils.capitalize(modelName) + Constants.JAVA_EXTENSION), ModelTemplate.getModel(modelName, packageName));
+        logger.lifecycle(PluginTemplate.WRITED_IN_FILES);
 
     }
 }
