+++ src/main/java/co/com/bancolombia/task/GenerateUseCaseTask.java
@@ -1,7 +1,9 @@
 package co.com.bancolombia.task;
 
-import co.com.bancolombia.Constants;
+import co.com.bancolombia.templates.Constants;
 import co.com.bancolombia.Utils;
+import co.com.bancolombia.templates.PluginTemplate;
+import co.com.bancolombia.templates.UseCaseTemplate;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.tasks.TaskAction;
@@ -27,13 +29,13 @@ public class GenerateUseCaseTask extends DefaultTask {
         String useCaseDir = Constants.DOMAIN.concat("/").concat(Constants.USECASE).concat("/").concat(Constants.MAIN_JAVA).concat("/").concat(packageName).concat("/").concat(Constants.USECASE).concat("/").concat(Utils.decapitalize(useCaseName));
         logger.lifecycle("Use Case Name: {}", useCaseName);
 
-        logger.lifecycle("Generating Childs Dirs");
+        logger.lifecycle(PluginTemplate.GENERATED_CHILDS_DIRS);
         getProject().mkdir(useCaseDir);
-        logger.lifecycle("Generated Childs Dirs");
+        logger.lifecycle(PluginTemplate.GENERATING_CHILDS_DIRS);
 
-        logger.lifecycle("Writing in Files");
-        Utils.writeString(getProject(), useCaseDir.concat("/").concat(Utils.capitalize(useCaseName) + Constants.JAVA_EXTENSION), Constants.getUseCase(useCaseName, packageName));
-        logger.lifecycle("Writed in Files");
+        logger.lifecycle(PluginTemplate.GENERATING_FILES);
+        Utils.writeString(getProject(), useCaseDir.concat("/").concat(Utils.capitalize(useCaseName) + Constants.JAVA_EXTENSION), UseCaseTemplate.getUseCase(useCaseName, packageName));
+        logger.lifecycle(PluginTemplate.WRITED_IN_FILES);
     }
 
     private void throwUseCase(){
