+++ src/main/java/co/com/bancolombia/task/ValidateStructureTask.java
@@ -1,6 +1,6 @@
 package co.com.bancolombia.task;
 
-import co.com.bancolombia.Constants;
+import co.com.bancolombia.templates.Constants;
 import co.com.bancolombia.Utils;
 import co.com.bancolombia.exceptions.CleanException;
 import org.gradle.api.DefaultTask;
@@ -60,7 +60,7 @@ public class ValidateStructureTask extends DefaultTask {
         Stream<String> stream = Utils.readFile(getProject(), Constants.DOMAIN.concat("/").concat(Constants.MODEL).concat("/").concat(Constants.BUILD_GRADLE));
 
         long countImplementationproject = stream
-                .map(line -> line.replaceAll(" ", ""))
+                .map(line -> line.replaceAll("\\s", ""))
                 .filter(line -> !line.startsWith("//") && line.contains("implementationproject"))
                 .count();
         return countImplementationproject == 0;
@@ -79,10 +79,10 @@ public class ValidateStructureTask extends DefaultTask {
                 return null;
             }
         };
-        boolean isvalid = stream.get().filter(line -> !line.startsWith("//")).map(line -> line.replaceAll(" ", ""))
+        boolean isvalid = stream.get().filter(line -> !line.startsWith("//")).map(line -> line.replaceAll("\\s", ""))
                 .anyMatch(str -> str.equals(modelDependency1) || str.equals(modelDependency2));
-        long countImplementationproject = stream.get().map(line -> line.replaceAll(" ", "")).filter(line -> !line.startsWith("//") && line.contains("implementationproject")).count();
-        long countCompileproject = stream.get().map(line -> line.replaceAll(" ", "")).filter(line -> !line.startsWith("//") && line.contains("compileproject")).count();
+        long countImplementationproject = stream.get().map(line -> line.replaceAll("\\s", "")).filter(line -> !line.startsWith("//") && line.contains("implementationproject")).count();
+        long countCompileproject = stream.get().map(line -> line.replaceAll("\\s", "")).filter(line -> !line.startsWith("//") && line.contains("compileproject")).count();
 
         return isvalid && ((countImplementationproject == 1 && countCompileproject == 0) || (countImplementationproject == 0 && countCompileproject == 1));
     }
