+++ src/main/java/co/com/bancolombia/templates/PluginTemplate.java
@@ -1,0 +1,12 @@
+package co.com.bancolombia.templates;
+
+public class PluginTemplate {
+
+    private PluginTemplate(){}
+
+    public static final String GENERATING_FILES = "Generating Files";
+    public static final String GENERATING_CHILDS_DIRS = "Generating Childs Dirs";
+    public static final String GENERATED_CHILDS_DIRS = "Childs Dirs Generated";
+    public static final String WRITED_IN_FILES = "Files Generated ";
+    public static final String VERSION_PLUGIN = "1.5.1";
+}
