+++ src/test/java/co/com/bancolombia/ConstantsTest.java
@@ -1,7 +1,7 @@
 package co.com.bancolombia;
 
+import co.com.bancolombia.templates.*;
 import org.junit.Assert;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
 public class ConstantsTest {
@@ -10,7 +10,7 @@ public class ConstantsTest {
     public void getSettingsGradleContent() {
         String nameProject = "testProjectName";
 
-        String act = Constants.getSettingsGradleContent(nameProject);
+        String act = ScaffoldTemplate.getSettingsGradleContent(nameProject);
 
         Assert.assertTrue(act.contains(nameProject));
         Assert.assertTrue(act instanceof  String);
@@ -21,7 +21,7 @@ public class ConstantsTest {
     public void getSettingsGradleContentException() {
         String nameProject = null;
 
-        String act = Constants.getSettingsGradleContent(nameProject);
+        String act = ScaffoldTemplate.getSettingsGradleContent(nameProject);
 
         Assert.assertTrue(act.contains(nameProject));
         Assert.assertTrue(act instanceof  String);
@@ -32,7 +32,7 @@ public class ConstantsTest {
     public void getApplicationPropertiesContent() {
         String nameProject = "testProjectName";
 
-        String act = Constants.getApplicationPropertiesContent(nameProject);
+        String act = ScaffoldTemplate.getApplicationPropertiesContent(nameProject);
 
         Assert.assertTrue(act.contains(nameProject));
         Assert.assertTrue(act instanceof  String);
@@ -42,7 +42,7 @@ public class ConstantsTest {
     public void getMainApplicationContent() {
         String nameProject = "testProjectName";
 
-        String act = Constants.getMainApplicationContent(nameProject);
+        String act = ScaffoldTemplate.getMainApplicationContent(nameProject);
 
         Assert.assertTrue(act.contains(nameProject));
         Assert.assertTrue(act instanceof  String);
@@ -54,7 +54,7 @@ public class ConstantsTest {
         String modelName = "testProjectName";
         String packageName = "demo.package";
 
-        String act = Constants.getModel(modelName, packageName);
+        String act = ModelTemplate.getModel(modelName, packageName);
 
         Assert.assertTrue(act.contains(modelName));
         Assert.assertTrue(act.contains(packageName));
@@ -64,7 +64,7 @@ public class ConstantsTest {
 
     @Test
     public void getBuildGradleContent() {
-        String act =  Constants.getBuildGradleContent();
+        String act =  ScaffoldTemplate.getBuildGradleContent();
 
         Assert.assertTrue(act instanceof  String);
     }
@@ -74,7 +74,7 @@ public class ConstantsTest {
         String modelName = "testProjectName";
         String packageName = "demo.package";
 
-        String act =  Constants.getInterfaceModel(modelName, packageName);
+        String act =  ModelTemplate.getInterfaceModel(modelName, packageName);
 
         Assert.assertTrue(act.contains(modelName));
         Assert.assertTrue(act.contains(packageName));
@@ -85,7 +85,7 @@ public class ConstantsTest {
     public void getGradlePropertiesContent() {
         String packageName = "demo.package";
 
-        String act =  Constants.getGradlePropertiesContent(packageName);
+        String act =  ScaffoldTemplate.getGradlePropertiesContent(packageName);
 
         Assert.assertTrue(act.contains(packageName));
         Assert.assertTrue(act instanceof  String);
@@ -96,7 +96,7 @@ public class ConstantsTest {
         String useCaseName = "testProjectName";
         String packageName = "demo.package";
 
-        String act =  Constants.getUseCase(useCaseName, packageName);
+        String act =  UseCaseTemplate.getUseCase(useCaseName, packageName);
 
         Assert.assertTrue(act.contains(useCaseName));
         Assert.assertTrue(act.contains(packageName));
@@ -108,7 +108,7 @@ public class ConstantsTest {
     public void getApiRestClassContent(){
         String packageName = "demo.package";
 
-        String act =  Constants.getApiRestClassContent(packageName);
+        String act =  EntryPointTemplate.getApiRestClassContent(packageName);
 
         Assert.assertTrue(act.contains(packageName));
         Assert.assertTrue(act instanceof  String);
@@ -116,35 +116,35 @@ public class ConstantsTest {
     }
     @Test
     public void getBuildGradleHelperMongoRepository(){
-        String act =  Constants.getBuildGradleHelperMongoRepository();
+        String act =  HelperTemplate.getBuildGradleHelperMongoRepository();
 
         Assert.assertTrue(act instanceof  String);
     }
 
     @Test
     public void getBuildGradleHelperJPARepository(){
-        String act =  Constants.getBuildGradleHelperJPARepository();
+        String act =  HelperTemplate.getBuildGradleHelperJPARepository();
 
         Assert.assertTrue(act instanceof  String);
     }
 
     @Test
     public void getBuildGradleMongoRepository(){
-        String act =  Constants.getBuildGradleMongoRepository();
+        String act =  DrivenAdapterTemplate.getBuildGradleMongoRepository();
 
         Assert.assertTrue(act instanceof  String);
     }
 
     @Test
     public void getBuildGradleSecretsManager(){
-        String act =  Constants.getBuildGradleSecretsManager();
+        String act =  DrivenAdapterTemplate.getBuildGradleSecretsManager();
 
         Assert.assertTrue(act instanceof  String);
     }
 
     @Test
     public void getBuildGradleApiRest(){
-        String act =  Constants.getBuildGradleApiRest();
+        String act =  EntryPointTemplate.getBuildGradleApiRest();
 
         Assert.assertTrue(act instanceof  String);
     }
